app/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==1.0.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Werkzeug==2.0.2
mysql-connector-python==8.0.28
flask-mysql==1.5.2